refactor: move TableChartSqlConfig to common
